feat: add ArtworkEditionSetItem component
